pyproject.toml
[project]
name = "sonika"
version = "0.2.1"
description = "Autonomous tool execution layer & CLI for AI orchestrators"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Sonika Team"}]

dependencies = [
    "langchain>=0.3",
    "langchain-core>=0.3",
    "pydantic>=2.0",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "sonika-ai-toolkit>=0.2.7",
    "langchain-google-genai",
    "langchain-aws",
    "prompt_toolkit>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]

[project.scripts]
sonika = "sonika.cli:app"

[tool.setuptools.packages.find]
include = ["sonika*"]
exclude = ["skills*", "tests*"]
